+++ pkg/postage/noop.go
@@ -47,8 +47,4 @@ func (b *NoOpBatchStore) StorageRadius() uint8 { return 0 }
 
 func (b *NoOpBatchStore) SetStorageRadius(func(uint8) uint8) error { return nil }
 
-func (b *NoOpBatchStore) SetStorageRadiusSetter(StorageRadiusSetter) {}
-
-func (b *NoOpBatchStore) Unreserve(UnreserveIteratorFn) error { return nil }
-
 func (b *NoOpBatchStore) Reset() error { return nil }
